Implement the matter service
This service communicates with the backend to get
the historic data. It also pases this data to adapt
it to what the components need, and in the process
gets rid of the snake casing found in some of the
data objects of the backend.

The components will not access the historical data
directly. They will access the matters data, which
is simply parsed historical data.