Apply basic filtering, simplify matterservice
There is no need to keep around the history object in
the application. The parsing to a matter array now
happens the moment the history data arrives to the
frontend.

No other piece of code knows now about the
history object, just the function that asks the
API for it.

Also, filtermatters works correctly, despite it
only filters by type at the moment.