Implement core changes in matters handling
Both the filtering and the paginating logic are now
being processed through the same function. The amount
of code used for this task has been notably reduced
due to the convergence of the matters parsing
logic in just one point. This has been made
possible thanks to correct observable piping.